requirements.txt
umap-learn
sentence_transformers
numpy